refactor(react): move `StoreRegistryContext` to root of package
