+++ packages/@livestore/react/src/experimental/multi-store/mod.ts
@@ -1,4 +1,3 @@
-export * from '../../StoreRegistryContext.tsx'
 export * from './StoreRegistry.ts'
 export * from './storeOptions.ts'
 export * from './types.ts'
