+++ packages/@livestore/react/src/mod.ts
@@ -1,5 +1,6 @@
 export { LiveStoreContext, type ReactApi } from './LiveStoreContext.ts'
 export { LiveStoreProvider } from './LiveStoreProvider.tsx'
+export * from './StoreRegistryContext.tsx'
 export {
   type Dispatch,
   type SetStateAction,
