+++ packages/@livestore/react/src/useStore.ts
@@ -1,9 +1,9 @@
 import type { LiveStoreSchema } from '@livestore/common/schema'
 import type { Store } from '@livestore/livestore'
 import React from 'react'
-import { useStoreRegistry } from './experimental/multi-store/StoreRegistryContext.tsx'
 import type { CachedStoreOptions } from './experimental/multi-store/types.ts'
 import type { ReactApi } from './LiveStoreContext.ts'
+import { useStoreRegistry } from './StoreRegistryContext.tsx'
 import { useClientDocument } from './useClientDocument.ts'
 import { useQuery } from './useQuery.ts'
